Fix : tanstack Query api 요청문제 해결 , Refactor : api/route folder 위치 수정, 기존의 zustand, tmdb.ts 삭제
