+++ netflix-assignment/src/hooks/useFetchContents.ts
@@ -2,20 +2,19 @@ import { useQuery } from '@tanstack/react-query';
 import axios from 'axios';
 import { Movie } from '@/types/movie';
 import { TvShow } from '@/types/tvshows';
-import instance from '@/utils/axiosInstance';
 
 const fetchMoviesByCategory = async (category: string): Promise<Movie[]> => {
   const response = await axios.get(`/movie/${category}`);
-  return response.data.results;
+  return response.data;
 };
 
 const fetchTvShowsByCategory = async (category: string): Promise<TvShow[]> => {
   const response = await axios.get(`/tv/${category}`);
-  return response.data.results;
+  return response.data;
 };
 
 // 모든 카테고리의 영화를 가져오는 함수
-const fetchAllMovies = async (): Promise<{ [key: string]: Movie[] }> => {
+const fetchAllMovies = async (): Promise<{ [key: string]: Movie[]}> => {
   const categories = ['popular', 'top_rated', 'now_playing', 'upcoming'];
   const moviePromises = categories.map(fetchMoviesByCategory);
   const results = await Promise.all(moviePromises);
@@ -27,7 +26,7 @@ const fetchAllMovies = async (): Promise<{ [key: string]: Movie[] }> => {
 };
 
 // 모든 카테고리의 TV 시리즈를 가져오는 함수
-const fetchAllTvShows = async (): Promise<{ [key: string]: TvShow[] }> => {
+const fetchAllTvShows = async (): Promise<{ [key: string]: TvShow[]}> => {
   const categories = ['popular', 'top_rated', 'airing_today', 'on_the_air'];
   const tvPromises = categories.map(fetchTvShowsByCategory);
   const results = await Promise.all(tvPromises);
@@ -46,33 +45,29 @@ export const useFetchMovies = (category: string) => {
     });
   };
   
-  // 개별 카테고리의 TV 시리즈를 가져오는 훅
-  export const useFetchTvShows = (category: string) => {
+// 개별 카테고리의 TV 시리즈를 가져오는 훅
+export const useFetchTvShows = (category: string) => {
     return useQuery({
-      queryKey: ['tv', category],
-      queryFn: () => fetchTvShowsByCategory(category),
-      staleTime: 1000 * 60 * 60,
+        queryKey: ['tv', category],
+        queryFn: () => fetchTvShowsByCategory(category),
+        staleTime: 1000 * 60 * 60,
     });
-  };
-  
-  // **모든 카테고리의 영화를 가져오는 훅**
-  export const useFetchAllMovies = () => {
+};
+
+// **모든 카테고리의 영화를 가져오는 훅**
+export const useFetchAllMovies = () => {
     return useQuery({
-      queryKey: ['allMovies'],
-      queryFn: fetchAllMovies,
-      staleTime: 1000 * 60 * 60,
+        queryKey: ['allMovies'],
+        queryFn: fetchAllMovies,
+        staleTime: 1000 * 60 * 60,
     });
-  };
-  
-  // **모든 카테고리의 TV 시리즈를 가져오는 훅**
-  export const useFetchAllTvShows = () => {
+};
+
+// **모든 카테고리의 TV 시리즈를 가져오는 훅**
+export const useFetchAllTvShows = () => {
     return useQuery({
-      queryKey: ['allTvShows'],
-      queryFn: fetchAllTvShows,
-      staleTime: 1000 * 60 * 60,
+        queryKey: ['allTvShows'],
+        queryFn: fetchAllTvShows,
+        staleTime: 1000 * 60 * 60,
     });
-  };
-  const fetchAllMovies2 = async () => {
-    const response = await axios.get('/api/movies');
-    return response.data;
-  };
+};
